[3568] Refactor code
