+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/message/list/adapter/viewholder/attachment/QuotedAttachmentFactoryManager.kt
@@ -1,3 +1,19 @@
+/*
+ * Copyright (c) 2014-2022 Stream.io Inc. All rights reserved.
+ *
+ * Licensed under the Stream License;
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    https://github.com/GetStream/stream-chat-android/blob/main/LICENSE
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package io.getstream.chat.android.ui.message.list.adapter.viewholder.attachment
 
 import android.view.ViewGroup
@@ -18,8 +34,8 @@ public class QuotedAttachmentFactoryManager(
      * @return True if there is a factory that can handle the attachments from this quoted message.
      */
     public fun canHandle(message: Message): Boolean {
-        return quotedAttachmentFactories.any { it.canHandle(message) }
-            || ChatUI.attachmentFactoryManager.canHandle(message)
+        return quotedAttachmentFactories.any { it.canHandle(message) } ||
+            ChatUI.attachmentFactoryManager.canHandle(message)
     }
 
     /**
